Move entities interfaces from game bundle into component
