+++ src/Badger/Bundle/GameBundle/Controller/QuestController.php
@@ -2,7 +2,7 @@
 
 namespace Badger\GameBundle\Controller;
 
-use Badger\GameBundle\Entity\QuestInterface;
+use Badger\Component\Game\Model\QuestInterface;
 use Badger\GameBundle\Form\QuestType;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\Form\Form;
