+++ src/Badger/Bundle/GameBundle/Doctrine/Repository/UnlockedBadgeRepository.php
@@ -2,7 +2,7 @@
 
 namespace Badger\GameBundle\Doctrine\Repository;
 
-use Badger\GameBundle\Entity\BadgeInterface;
+use Badger\Component\Game\Model\BadgeInterface;
 use Badger\Component\Game\Repository\TagSearchableRepositoryInterface;
 use Badger\UserBundle\Entity\UserInterface;
 use Doctrine\DBAL\Connection;
