+++ src/Badger/Bundle/GameBundle/Entity/Adventure.php
@@ -2,6 +2,9 @@
 
 namespace Badger\GameBundle\Entity;
 
+use Badger\Component\Game\Model\AdventureInterface;
+use Badger\Component\Game\Model\AdventureStepInterface;
+use Badger\Component\Game\Model\BadgeInterface;
 use Badger\TagBundle\Entity\TagInterface;
 use Doctrine\Common\Collections\ArrayCollection;
