+++ src/Badger/Bundle/GameBundle/Entity/ClaimedBadge.php
@@ -2,6 +2,8 @@
 
 namespace Badger\GameBundle\Entity;
 
+use Badger\Component\Game\Model\BadgeInterface;
+use Badger\Component\Game\Model\ClaimedBadgeInterface;
 use Badger\UserBundle\Entity\User;
 
 /**
