+++ src/Badger/Bundle/GameBundle/Entity/Quest.php
@@ -2,6 +2,7 @@
 
 namespace Badger\GameBundle\Entity;
 
+use Badger\Component\Game\Model\QuestInterface;
 use Badger\TagBundle\Entity\TagInterface;
 use Doctrine\Common\Collections\ArrayCollection;
