+++ src/Badger/Bundle/GameBundle/Factory/ClaimedBadgeFactory.php
@@ -3,7 +3,7 @@
 namespace Badger\GameBundle\Factory;
 
 use Badger\Component\Game\Factory\ClaimedBadgeFactoryInterface;
-use Badger\GameBundle\Entity\BadgeInterface;
+use Badger\Component\Game\Model\BadgeInterface;
 use Badger\GameBundle\Entity\ClaimedBadge;
 use Badger\UserBundle\Entity\UserInterface;
