+++ src/Badger/Bundle/GameBundle/Unlocker/BadgeUnlocker.php
@@ -3,8 +3,8 @@
 namespace Badger\GameBundle\Unlocker;
 
 use Badger\Component\Game\Unlocker\BadgeUnlockerInterface;
-use Badger\GameBundle\Entity\BadgeInterface;
-use Badger\GameBundle\Entity\ClaimedBadgeInterface;
+use Badger\Component\Game\Model\BadgeInterface;
+use Badger\Component\Game\Model\ClaimedBadgeInterface;
 use Badger\Component\Game\Factory\UnlockedBadgeFactoryInterface;
 use Badger\Component\Game\Repository\UnlockedBadgeRepositoryInterface;
 use Badger\StorageUtilsBundle\Remover\RemoverInterface;
