+++ src/Badger/Component/Game/Factory/ClaimedBadgeFactoryInterface.php
@@ -2,8 +2,8 @@
 
 namespace Badger\Component\Game\Factory;
 
-use Badger\GameBundle\Entity\BadgeInterface;
-use Badger\GameBundle\Entity\ClaimedBadgeInterface;
+use Badger\Component\Game\Model\BadgeInterface;
+use Badger\Component\Game\Model\ClaimedBadgeInterface;
 use Badger\UserBundle\Entity\UserInterface;
 
 /**
