+++ src/Badger/Component/Game/Model/QuestInterface.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace Badger\GameBundle\Entity;
+namespace Badger\Component\Game\Model;
 
+use Badger\GameBundle\Entity\Quest;
 use Badger\TagBundle\Taggable\TaggableInterface;
 use Doctrine\Common\Collections\ArrayCollection;
